TestMiniAPP/requirements.txt
aiogram==3.10.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.0.3
openpyxl==3.1.2
python-dotenv==1.0.0
Jinja2==3.1.2
aiofiles==23.2.1
python-multipart==0.0.6
typing-extensions==4.8.0